Refactoring: better files
