+++ src/devices/plethora.rs
@@ -1,4 +1,6 @@
-use crate::devices::{Direction, MidiMessages};
+use crate::devices::Direction;
+use crate::handler::MidiMessages;
+
 use midi_types::{Channel, Control, MidiMessage, Program, Value7};
 
 const PLETHORA_CHANNEL: Channel = Channel::new(1);
